+++ main.js
@@ -1,0 +1,28 @@
+import React from 'react';
+import { AppRegistry } from 'react-native';
+import AppContainer from './app/containers/AppContainer';
+import { Provider } from 'react-redux';
+import { createStore, applyMiddleware, combineReducers, compose } from 'redux';
+import thunkMiddleware from 'redux-thunk';
+import { createLogger } from 'redux-logger';
+import reducer from './app/reducers';
+
+const loggerMiddleware = createLogger({ predicate: (getState, action) => __DEV__  });
+
+function configureStore(initialState) {
+  const enhancer = compose(
+    applyMiddleware(
+      thunkMiddleware,
+      loggerMiddleware
+    )
+  );
+  return createStore(reducer, initialState, enhancer);
+}
+
+const store = configureStore({});
+
+export const App = () => (
+  <Provider store={store}>
+    <AppContainer />
+  </Provider>
+);
